requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
playwright==1.55.0
playwright-stealth==2.0.0
SQLAlchemy==2.0.44
beautifulsoup4==4.14.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
